requirements.txt
Django>=3.0.5
djangorestframework>=3.11.0
django-rest-swagger>=2.2.0
request>=2019.4.13
ipdata>=3.2
psycopg2>=2.8.5
psycopg2-binary>=2.8.5
django-cors-headers>=3.2.1
Pillow>=7.1.1
django-admin-rangefilter>=0.5.4
django-filter>=2.2.0
drf-yasg>=1.17.1

python-decouple>=3.3
drf-yasg>=1.17.1
django-auditlog>=0.4.7
PyJWT>=1.7.1
celery>=4.4.2
scipy